+++ insanic/authentication/__init__.py
@@ -165,7 +165,14 @@ class ServiceJWTAuthentication(BaseJSONWebTokenAuthentication):
         return handlers.jwt_service_decode_handler(token)
 
     async def authenticate_credentials(self, request, payload):
-        user = User(**payload.pop('user', AnonymousUser))
+        user_params = {"id": "", "level": -1}
+
+        for f in request.headers.get(settings.INTERNAL_REQUEST_USER_HEADER, '').split(';'):
+            if f:
+                k, v = f.split('=')
+                user_params.update({k: v})
+
+        user = User(**user_params)
 
         service = RequestService(is_authenticated=True, **payload)