+++ insanic/services/__init__.py
@@ -15,7 +15,7 @@ from insanic.authentication.handlers import jwt_service_encode_handler, jwt_serv
 from insanic.conf import settings
 from insanic.errors import GlobalErrorCodes
 from insanic.functional import cached_property_with_ttl
-from insanic.models import AnonymousUser
+from insanic.models import AnonymousUser, to_header_value
 from insanic.services.response import InsanicResponse
 from insanic.scopes import is_docker
 from insanic.tracing.clients import aws_xray_trace_config
@@ -28,13 +28,22 @@ from insanic.utils.datetime import get_utc_datetime
 DEFAULT_SERVICE_REQUEST_TIMEOUT = 1
 
 
-def service_auth_token(service):
+# def service_auth_token(service):
+#     default = dict(AnonymousUser)
+#     try:
+#         user = aiotask_context.get(settings.TASK_CONTEXT_REQUEST_USER, default)
+#     except AttributeError:
+#         user = default
+#     return jwt_service_encode_handler(jwt_service_payload_handler(service, user))
+
+
+def user_context():
     default = dict(AnonymousUser)
     try:
         user = aiotask_context.get(settings.TASK_CONTEXT_REQUEST_USER, default)
     except AttributeError:
         user = default
-    return jwt_service_encode_handler(jwt_service_payload_handler(service, user))
+    return user
 
 
 class ServiceRegistry(dict):
@@ -71,20 +80,18 @@ class ServiceRegistry(dict):
 
 class Service:
     _session = None
+    _semaphore = None
 
     def __init__(self, service_type):
-        self._registry = ServiceRegistry()
         self._service_name = service_type
         # self._session = None
+        self._token = None
 
-    # @property
-    # def service_auth_token(self):
-    #     default = dict(AnonymousUser)
-    #     try:
-    #         user = aiotask_context.get(settings.TASK_CONTEXT_REQUEST_USER, default)
-    #     except AttributeError:
-    #         user = default
-    #     return jwt_service_encode_handler(jwt_service_payload_handler(self, user))
+    @classmethod
+    def semaphore(cls):
+        if cls._semaphore is None:
+            cls._semaphore = asyncio.Semaphore(20)
+        return cls._semaphore
 
     @property
     def service_name(self):
@@ -101,6 +108,7 @@ class Service:
 
     @cached_property_with_ttl(ttl=60)
     def host(self):
+        # return "localhost"
         _host = settings.SERVICE_GLOBAL_HOST_TEMPLATE.format(self.service_name)
         if hasattr(settings, "SWARM_SERVICE_LIST"):
             _host = settings.SWARM_SERVICE_LIST.get(self.service_name, {}).get('host', _host)
@@ -116,26 +124,16 @@ class Service:
         return _port
 
     @property
+    def service_token(self):
+        if self._token is None:
+            self._token = jwt_service_encode_handler(jwt_service_payload_handler(self))
+
+        return self._token
+
+    @property
     def url(self):
         url_partial_path = "/api/v1/"
         return URL(f"{self.schema}://{self.host}:{self.port}{url_partial_path}")
-
-    # @property
-    # def session(self):
-    #     if self._session is None or self._session.loop.is_closed() is True:
-    #
-    #         trace_configs = None
-    #
-    #         if settings.TRACING_ENABLED:
-    #             trace_configs = [aws_xray_trace_config(tracing_name(self.service_name))]
-    #
-    #         self._session = aiohttp.ClientSession(loop=get_event_loop(),
-    #                                               connector=aiohttp.TCPConnector(limit_per_host=10,
-    #                                                                              ttl_dns_cache=300),
-    #                                               response_class=InsanicResponse,
-    #                                               read_timeout=DEFAULT_SERVICE_REQUEST_TIMEOUT,
-    #                                               trace_configs=trace_configs)
-    #     return self._session
 
     @classmethod
     def session(cls):
@@ -148,16 +146,15 @@ class Service:
                 trace_configs = [aws_xray_trace_config()]
 
             cls._session = aiohttp.ClientSession(loop=get_event_loop(),
-                                                 connector=aiohttp.TCPConnector(limit=200,
-                                                                                limit_per_host=50,
-                                                                                 ttl_dns_cache=300),
-                                                  response_class=InsanicResponse,
-                                                  read_timeout=DEFAULT_SERVICE_REQUEST_TIMEOUT,
-                                                  trace_configs=trace_configs)
+                                                 connector=aiohttp.TCPConnector(limit=100,
+                                                                                keepalive_timeout=15,
+                                                                                limit_per_host=20,
+                                                                                ttl_dns_cache=10),
+                                                 response_class=InsanicResponse,
+                                                 read_timeout=DEFAULT_SERVICE_REQUEST_TIMEOUT,
+                                                 trace_configs=trace_configs)
 
         return cls._session
-
-
 
     def _construct_url(self, endpoint, query_params={}):
         url = self.url.with_path(endpoint)
@@ -165,13 +162,35 @@ class Service:
             url = url.with_query(**query_params)
         return url
 
-    @capture_async("insanic_http_dispatch")
-    async def http_dispatch(self, method, endpoint, req_ctx=None, *, query_params=None, payload=None,
+    async def http_dispatch(self, method, endpoint, *, query_params=None, payload=None,
                             files=None, headers=None,
                             propagate_error=False, skip_breaker=False, include_status_code=False, request_timeout=None):
-        if req_ctx is not None:
-            warnings.warn("`req_ctx`, the 3rd argument is deprecated.  Please remove. "
-                          "Sending this will do absolutely nothing.")
+        """
+        Interface for sending requests to other services.
+
+        :param method: method to send request (GET, POST, PATCH, PUT, etc)
+        :type method: string
+        :param endpoint: the path to send request to (eg /api/v1/..)
+        :type endpoint: string
+        :param query_params: query params to attach to url
+        :type query_params: dict
+        :param payload: the data to send on any non GET requests
+        :type payload: dict
+        :param files: if any files to send with request, must be included here
+        :type files: dict
+        :param headers: headers to send along with request
+        :type headers: dict
+        :param propagate_error: if you want to raise on 400 or greater status codes
+        :type propagate_error: bool
+        :param skip_breaker: if you want to skip circuit breaker
+        :type skip_breaker: bool
+        :param include_status_code: if you want this method to return the response with the status code
+        :type include_status_code: bool
+        :param request_timeout: if you want to increase the timeout for this requests
+        :type request_timeout: int
+        :return:
+        :rtype: dict or tuple(dict, int)
+        """
 
         files = files or {}
         query_params = query_params or {}
@@ -192,6 +211,20 @@ class Service:
             return response, status_code
         return response
 
+    def prepare_request(self, method, endpoint, *, query_params=None, payload=None,
+                        files=None, headers=None):
+        files = files or {}
+        query_params = query_params or {}
+        payload = payload or {}
+        headers = headers or {}
+        if method.upper() not in HTTP_METHODS:
+            raise ValueError("{0} is not a valid method.".format(method))
+
+        url = self._construct_url(endpoint, query_params=query_params)
+        headers = self._prepare_headers(headers, files)
+        data = self._prepare_body(headers, payload, files)
+
+        return {"method": method, "url": url, "headers": headers, "data": data, "service_name": self.service_name}
 
     def _prepare_headers(self, headers, files=None):
 
@@ -208,8 +241,10 @@ class Service:
         lower_headers.update({"date": get_utc_datetime().strftime("%a, %d %b %y %T %z")})
 
         lower_headers.update(
-            {"authorization": f"{settings.JWT_SERVICE_AUTH['JWT_AUTH_HEADER_PREFIX']} {service_auth_token(self)}"})
+            {"authorization": f"{settings.JWT_SERVICE_AUTH['JWT_AUTH_HEADER_PREFIX']} {self.service_token}"})
 
+        context_user = user_context()
+        lower_headers.update({f"{settings.INTERNAL_REQUEST_USER_HEADER.lower()}": to_header_value(context_user)})
         return lower_headers
 
 
@@ -220,7 +255,11 @@ class Service:
         if len(files) == 0 and headers.get('content-type') == "application/json":
             data = aiohttp.payload.JsonPayload(payload)
         else:
-            data = aiohttp.formdata.FormData(payload)
+            data = aiohttp.formdata.FormData()
+            for k, v in payload.items():
+                data.add_field(k, v, content_type=None)
+
+
             for k, v in files.items():
                 if k in payload.keys():
                     raise RuntimeError(f"CONFLICT ERROR: payload already has the key, {k}. Can not overwrite.")
@@ -235,8 +274,8 @@ class Service:
 
         return data
 
-    @capture_async("insanic__dispatch_fetch")
-    async def _dispatch_fetch(self, method, url, headers, data, **kwargs):
+    @classmethod
+    async def _dispatch_fetch(cls, method, url, headers, data, service_name, **kwargs):
 
         request_params = {
             "method": method,
@@ -245,22 +284,17 @@ class Service:
             "data": data
         }
 
-        timeout = kwargs.pop('request_timeout')
+        timeout = kwargs.pop('request_timeout', None)
         if timeout:
             request_params.update({"timeout": timeout})
 
-        request_params.update({"trace_request_ctx": {"name": tracing_name(self.service_name)}})
+        request_params.update({"trace_request_ctx": {"name": tracing_name(service_name)}})
 
-        async with self.session().request(**request_params) as resp:
-            await resp.read()
-            return resp
+        async with cls.semaphore():
+            async with cls.session().request(**request_params) as resp:
+                await resp.read()
+                return resp
 
-    # def create_request_object(self, method, url, query_params, headers, data):
-    #     return aiohttp.ClientRequest(method, url,
-    #                                  params=query_params, headers=headers, data=data)
-
-
-    @capture_async("insanic__dispatch")
     async def _dispatch(self, method, endpoint, *, query_params, payload, files, headers,
                         propagate_error=False, skip_breaker=False, request_timeout=None):
         """
@@ -282,10 +316,12 @@ class Service:
         data = self._prepare_body(headers, payload, files)
 
         # outbound_request = self.create_request_object(method, url, query_params, headers, data)
-
+        exc = None
         try:
-            _response_obj = await self._dispatch_fetch(method, url, headers, data, skip_breaker=skip_breaker,
-                                                       request_timeout=request_timeout)
+            _response_obj = await self._dispatch_fetch(method, url, headers, data,
+                                                       skip_breaker=skip_breaker,
+                                                       request_timeout=request_timeout,
+                                                       service_name=self.service_name)
 
             if propagate_error:
                 _response_obj.raise_for_status()
@@ -311,59 +347,59 @@ class Service:
                                           error_code=response.get('error_code', error_code),
                                           status_code=status_code)
             exc.message = response.get('message', GlobalErrorCodes.unknown_error.name)
-            raise exc
+            # raise exc
+        except aiohttp.client_exceptions.ClientConnectorError as e:
+            """subset of connection errors that are initiated by an OSError exception"""
+            if e.errno == 61:
+                msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(self._service_name)
+                exc = exceptions.ServiceUnavailable503Error(description=msg,
+                                                            error_code=GlobalErrorCodes.service_unavailable,
+                                                            status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
+            else:
+                exc = exceptions.APIException(description=e.strerror,
+                                              error_code=GlobalErrorCodes.unknown_error,
+                                              status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        except aiohttp.client_exceptions.ServerFingerprintMismatch:
+            """server fingerprint mismatch"""
+            msg = "Server Fingerprint Mismatch"
+            exc = exceptions.APIException(description=msg,
+                                          error_code=GlobalErrorCodes.server_signature_error,
+                                          status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+        except aiohttp.client_exceptions.ServerConnectionError as e:
+            """ server connection related errors """
+            if isinstance(e, aiohttp.client_exceptions.ServerDisconnectedError):
+                """ server disconnected """
+                exc = exceptions.ServiceUnavailable503Error(description=e.message,
+                                                            error_code=GlobalErrorCodes.service_unavailable,
+                                                            status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
+            elif isinstance(e, aiohttp.client_exceptions.ServerTimeoutError):
+                """server operation timeout, (read timeout, etc)"""
+                exc = exceptions.APIException(description=e.args[0],
+                                              error_code=GlobalErrorCodes.service_timeout,
+                                              status_code=status.HTTP_504_GATEWAY_TIMEOUT)
+            else:
+                msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(self._service_name)
+                exc = exceptions.ServiceUnavailable503Error(description=msg,
+                                                            error_code=GlobalErrorCodes.service_unavailable,
+                                                            status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
 
         except aiohttp.client_exceptions.ClientConnectionError as e:
             # https://aiohttp.readthedocs.io/en/v3.0.1/client_reference.html#hierarchy-of-exceptions
-            if isinstance(e, aiohttp.client_exceptions.ClientConnectorError):
-                """subset of connection errors that are initiated by an OSError exception"""
-                if e.errno == 61:
-                    msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(self._service_name)
-                    exc = exceptions.ServiceUnavailable503Error(description=msg,
-                                                                error_code=GlobalErrorCodes.service_unavailable,
-                                                                status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
-                else:
-                    exc = exceptions.APIException(description=e.strerror,
-                                                  error_code=GlobalErrorCodes.unknown_error,
-                                                  status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
-            elif isinstance(e, aiohttp.client_exceptions.ServerConnectionError):
-                """ server connection related errors """
+            exc = exceptions.APIException(description=e.args[0],
+                                          error_code=GlobalErrorCodes.unknown_error,
+                                          status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-                if isinstance(e, aiohttp.client_exceptions.ServerDisconnectedError):
-                    """ server disconnected """
-                    exc = exceptions.ServiceUnavailable503Error(description=e.message,
-                                                                error_code=GlobalErrorCodes.service_unavailable,
-                                                                status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
-                elif isinstance(e, aiohttp.client_exceptions.ServerTimeoutError):
-                    """server operation timeout, (read timeout, etc)"""
-                    exc = exceptions.APIException(description=e.args[0],
-                                                  error_code=GlobalErrorCodes.service_timeout,
-                                                  status_code=status.HTTP_504_GATEWAY_TIMEOUT)
-                else:
-                    msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(self._service_name)
-                    exc = exceptions.ServiceUnavailable503Error(description=msg,
-                                                                error_code=GlobalErrorCodes.service_unavailable,
-                                                                status_code=status.HTTP_503_SERVICE_UNAVAILABLE)
-            elif isinstance(e, aiohttp.client_exceptions.ServerFingerprintMismatch):
-                """server fingerprint mismatch"""
-                msg = "Server Fingerprint Mismatch"
-                exc = exceptions.APIException(description=msg,
-                                              error_code=GlobalErrorCodes.server_signature_error,
-                                              status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
-            else:
-                exc = exceptions.APIException(description=e.args[0],
-                                              error_code=GlobalErrorCodes.unknown_error,
-                                              status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-            raise exc
         except asyncio.TimeoutError:
             exc = exceptions.RequestTimeoutError(description=f'Request to {self.service_name} took too long!',
                                                  error_code=GlobalErrorCodes.request_timeout,
                                                  status_code=status.HTTP_408_REQUEST_TIMEOUT)
-            raise exc
         except aiohttp.client_exceptions.ClientPayloadError as e:
-            raise
+            exc = e
         except aiohttp.client_exceptions.InvalidURL as e:
-            raise
-
-        return response, response_status
+            exc = e
+        else:
+            return response, response_status
+        finally:
+            if exc:
+                raise exc