+++ tests/test_models.py
@@ -18,7 +18,7 @@ def test_anonymous_user():
     anon_user = AnonymousUser
 
     assert anon_user.is_staff is False
-    assert anon_user.is_authenticated is False
+    assert anon_user.is_authenticated is 0
     assert anon_user.is_active is False
     assert anon_user.is_banned is False
 
@@ -29,7 +29,7 @@ def test_user_active(test_user):
     assert test_user.id == TEST_USER_ID
     assert test_user.level == UserLevels.ACTIVE
     assert test_user.is_staff is False
-    assert test_user.is_authenticated is True
+    assert test_user.is_authenticated is 1
     assert test_user.is_active is True
     assert test_user.is_banned is False
 
@@ -41,7 +41,7 @@ def test_user_ban(test_user):
     test_user.level = UserLevels.BANNED
 
     assert test_user.is_staff is False
-    assert test_user.is_authenticated is False
+    assert test_user.is_authenticated is 0
     assert test_user.is_active is False
     assert test_user.is_banned is True
 
@@ -52,7 +52,7 @@ def test_user_deactivated(test_user):
     test_user.level = UserLevels.DEACTIVATED
 
     assert test_user.is_staff is False
-    assert test_user.is_authenticated is False
+    assert test_user.is_authenticated is 0
     assert test_user.is_active is False
     assert test_user.is_banned is False
 
@@ -63,7 +63,7 @@ def test_user_staff(test_user):
     test_user.level = UserLevels.STAFF
 
     assert test_user.is_staff is True
-    assert test_user.is_authenticated is True
+    assert test_user.is_authenticated is 1
     assert test_user.is_active is True
     assert test_user.is_banned is False